+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 0.1.4 - 2017-05-12
+
+### Added
+- GoogleSearchBot in ita industries project
+
 ## 0.1.3 - 2017-04-27
 
 ### Added